feat: move session files to the session folder
