+++ ui/src/features/session/components/session-execution.tsx
@@ -37,6 +37,7 @@ import Underline from '@tiptap/extension-underline';
 import { useEditor } from '@tiptap/react';
 import StarterKit from '@tiptap/starter-kit';
 import { useEffect, useState } from 'react';
+import { Pomodoro } from './pomodoro';
 
 const INITIAL_CHECKLIST = [
   { id: 1, label: 'Text here', checked: true },
@@ -284,12 +285,10 @@ export default function SessionExecution() {
         <Grid.Col span={{ base: 12, md: 4 }}>
           <Stack gap="lg">
             {/* Pomodoro */}
-            <Card radius="md" withBorder padding="xl">
-              <Stack align="center" gap="lg">
-                <Title order={5}>Pomodoro</Title>
-                <p>pomodoro</p>
-              </Stack>
-            </Card>
+            <Pomodoro
+              sessionDuration={studySession.duration}
+              pomodoro={studySession.pomodoro}
+            />
 
             {/* Checklist */}
             <Card radius="md" withBorder padding="lg">
