+++ ui/src/routes/sessions/$id.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
 import { fetchMe } from '@/features/auth/api';
-import SessionExecution from '@/features/goals/components/session-execution';
+import SessionExecution from '@/features/session/components/session-execution';
 
 export const Route = createFileRoute('/sessions/$id')({
   beforeLoad: async () => {
